pyproject.toml
[project]
name = "agents-shorts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "requests>=2.32.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "whisper-openai>=20231117",
  "yt-dlp>=2023.12.30",
  "mcp>=1.0.0",
  "fastmcp>=0.1.0",
  "smolagents>=0.1.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0"
]
